pyproject.toml
[project]
name = "product-search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "beautifulsoup4",
    "selenium",
    "apscheduler",
    "rapidfuzz",
    "spacy",
    "httpx",
    "aiosmtplib",
    "jinja2",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "pytest",
    "pytest-cov",
    "webdriver-manager",
    "pytest-asyncio",
    "selenium-stealth",
    "apscheduler",
    "websockets",
    "groq",
    "email_validator"
]

